refactor: extract pure provision helpers to provisionHelpers.ts, fix isConnectionError timeout retry bug
